GH-438: Inline the proto archive extractor
